requirements.txt
clickhouse-connect
fastapi
uvicorn[standard]
pydantic
openai>=1.0.0
python-dotenv
